LPS-87059 Move Java class from portal-impl to impl module
